requirements.txt
# Core RAG dependencies
llama-index==0.10.12
llama-index-llms-ollama==0.1.3
llama-index-embeddings-huggingface==0.1.4
llama-index-vector-stores-faiss==0.1.2
llama-index-readers-file==0.1.6

# Vector store and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.5.1
torch==2.2.0
transformers==4.38.0

# Reranker
FlagEmbedding==1.2.3

# API and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Utilities
aiofiles==23.2.1
python-dotenv==1.0.0
loguru==0.7.2

# Storage
sqlalchemy==2.0.25

